refactor: ♻️ ゲームの更新にかかわるフックを分離し、プレイヤー情報の取得をuseGameから行うように変更
