+++ src/components/buttons/TileButton.tsx
@@ -1,11 +1,10 @@
 import circleImageUrl from "@/assets/circle.png";
 import crossImageUrl from "@/assets/cross.png";
 import { useGame } from "@/game/hooks/useGame";
+import { useUpdateGame } from "@/game/hooks/useUpdateGame";
 import { canClickTile, getTileState } from "@/game/logics/tileLogic";
-import { playersStateAtom } from "@/game/stores/atoms";
 import type { Tile } from "@/game/types/tile";
 import { cn } from "@/lib/utils";
-import { useAtom } from "jotai";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
 
@@ -13,8 +12,8 @@ export interface TileButtonProps {
   tile: Tile;
 }
 export const TileButton: FC<TileButtonProps> = ({ tile }) => {
-  const { currentTurn, board, gameOver, updateGameAndBoard } = useGame();
-  const [players] = useAtom(playersStateAtom);
+  const { currentTurn, board, gameOver, players } = useGame();
+  const { updateGameAndBoard } = useUpdateGame();
   const { t } = useTranslation();
 
   const state = getTileState(currentTurn, board.size, tile);
