+++ src/components/texts/ResultLogText.tsx
@@ -1,6 +1,5 @@
-import { playersStateAtom } from "@/game/stores/atoms";
+import { useGame } from "@/game/hooks/useGame";
 import type { GameResult } from "@/game/types/result";
-import { useAtom } from "jotai";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
 import { NeonText } from "../ui/NeonText";
@@ -9,7 +8,7 @@ export interface ResultLogTextProps {
   result: GameResult;
 }
 export const ResultLogText: FC<ResultLogTextProps> = ({ result }) => {
-  const [players] = useAtom(playersStateAtom);
+  const { players } = useGame();
   const { t } = useTranslation();
 
   return (
